\name{NEff-package}
\alias{NEff-package}
\alias{NEff}
\docType{package}
\title{ Calculating Effective Sizes Based on Known Demographic Parameters of a Population }
\description{
Effective population sizes (often abbreviated as "Neff") are essential in biodiversity monitoring and conservation. For the first time, calculating effective sizes with data obtained within less than a generation but considering demographic parameters is possible. This individual based model uses demographic parameters of a population to calculate annual effective sizes and effective population sizes (per generation). A defined number of alleles and loci will be used to simulate the genotypes of the individuals. Stepwise mutation rates can be included. Variations in life history parameters (sex ratio, sex-specific survival, recruitment rate, reproductive skew) are possible. These results will help managers to define existing populations as viable or not.  }
\details{
\tabular{ll}{
Package: \tab NEff\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-05-04\cr
License: \tab GPL version 2.0\cr
}
The package is based on function population. This function creates an individual based population model simulating genetics via a specific number of loci holding a specific number of alleles. Demography is given as input values and can be adapted to every life history parameter combination.
}
\author{
Annegret Grimm, Bernd Gruber & Klaus Henle

Maintainer: Annegret Grimm <annegret.grimm@ufz.de>
}
\references{
Grimm, A., B. Gruber, M. Hoehn, K. Enders, and K. Henle. submitted to Methods in Ecology and Evolution. A model-derived short-term estimation method of effective population size for species with overlapping generations.
}
\keyword{ effective population size }
\keyword{ annual effective size }
\keyword{ individual based model }
\keyword{ heterozygosity }
\keyword{ demography }
\keyword{ life history }
\keyword{ step-wise mutation }
\keyword{ population development }
\keyword{ population growth }
