\name{NEff-package}
\alias{NEff-package}
\alias{NEff}
\docType{package}
\title{ Calculating effective sizes based on known demographic parameters of a population }
\description{
This package uses demographic parameter of a population to calculate annual effective sizes and effective population sizes (per generation). It assumes that no migration and mutation is occuring. A defined number of alleles and loci will be used to simulate the genotypes of the individuals. These results will help managers to define existing populations as viable or not. }
\details{
\tabular{ll}{
Package: \tab NEff\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-30\cr
License: \tab GPL version 3.0\cr
}
The package is based on function population. This function creates an individual based population model simulating genetics via a specific number of loci holding a specific number of alleles. Demography is given as input values and will be stable across time.
}
\author{
Annegret Grimm, Bernd Gruber & Klaus Henle

Maintainer: Annegret Grimm <annegret.grimm@ufz.de>
}
\references{
Grimm, A., B. Gruber, M. Hoehn, K. Enders, and K. Henle. submitted to Molecular Ecology Resources. Assessing effective population size in vertebrate populations under different degrees of generational overlap.
}
\keyword{ effective population size }
\keyword{ annual effective size }
\keyword{ individual based model }
\keyword{ heterozygosity }
\keyword{ demography }
